requirements.txt
boto3==1.34.101
deepface==0.0.84
hdbscan==0.8.33
imutils==0.5.4
insightface==0.6.2
matplotlib==3.7.0
mediapipe==0.10.9
moviepy==1.0.3
mtcnn==0.1.1
numpy==1.23.1
numpy==1.23.5
onnxruntime==1.17.3
onnxruntime_gpu==1.12.0
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
pandas==1.5.3
Pillow==10.1.0
Pillow==9.4.0
Pillow==10.3.0
pydub==0.25.1
PyYAML==6.0
PyYAML==6.0.1
retina_face==0.0.14
scenedetect==0.6.2
scikit_learn==1.2.1
torch==1.13.1
torch==1.12.1
torchvision==0.14.1
tqdm==4.64.1
umap==0.1.1
umap_learn==0.5.6
wandb==0.16.0
webrtcvad==2.0.10
youtube_search_python==1.6.6
yt_dlp==2023.12.30
